requirements.txt
Flask==2.3.3
Flask-Session==0.5.0
mysql-connector-python==8.1.0
requests==2.31.0
python-dotenv==1.0.0
groq==0.4.1
